api-server/requirements.txt
wagtail>=6.4,<7.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.28,<1.0
django-filter>=24.0,<25.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
django-cors-headers>=4.6,<5.0
